- Move rate filters to new Utilities file - Deprecate, redirect legacy rate filtering methods in Calculations
