+++ EasyPost.Tests/ServicesTests/Beta/RateServiceTest.cs
@@ -1,9 +1,9 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using EasyPost.Calculation;
 using EasyPost.Models.API.Beta;
 using EasyPost.Tests._Utilities;
 using EasyPost.Tests._Utilities.Annotations;
+using EasyPost.Utilities;
 using EasyPost.Utilities.Annotations;
 using Xunit;
